Updating the readme and adding another test.
